backend/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
pylint-django==2.0.15
